+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/response/mapping/submodel/MapperHelper.java
@@ -43,6 +43,14 @@ public class MapperHelper {
                 )).get(globalAssetId);
     }
 
+    public static String getContractAgreementId(List<Shell> shells, String globalAssetId) {
+        return shells.stream()
+                .filter(shell -> shell.payload().globalAssetId().equals(globalAssetId))
+                .map(Shell::contractAgreementId)
+                .findFirst()
+                .orElse(null);
+    }
+
     public static OffsetDateTime getOffsetDateTime(String date) {
         try {
             return OffsetDateTime.parse(date);