+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/assets/infrastructure/base/IrsCallbackControllerIT.java
@@ -52,8 +52,9 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
 
 
     @Test
-    void givenNoAssets_whenCallbackReceived_thenSaveThem() {
+    void givenNoAssets_whenCallbackReceived_thenSaveThem() throws JoseException {
         // given
+
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsApiReturnsJobDetails();
         String jobId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
@@ -75,6 +76,18 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
         assertThat(bpnSupportRepository.findAll()).hasSize(6);
         assetsSupport.assertAssetAsBuiltSize(16);
         assetsSupport.assertAssetAsPlannedSize(0);
+        String contractAgreementId = given()
+                .header(oAuth2Support.jwtAuthorization(JwtRole.ADMIN))
+                .contentType(ContentType.JSON)
+                .log().all()
+                .when()
+                .pathParam("assetId", "urn:uuid:d387fa8e-603c-42bd-98c3-4d87fef8d2bb")
+                .get("/api/assets/as-built/{assetId}")
+                .then()
+                .log().all()
+                .statusCode(200)
+                .extract().path("contractAgreementId");
+        assertThat(contractAgreementId).isNotEmpty();
     }
 
     @Test