+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/common/support/EdcSupport.java
@@ -24,11 +24,14 @@ import org.glassfish.grizzly.http.util.HttpStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.util.regex.Pattern;
+
 import static com.xebialabs.restito.builder.stub.StubHttp.whenHttp;
 import static com.xebialabs.restito.builder.verify.VerifyHttp.verifyHttp;
 import static com.xebialabs.restito.semantics.Action.noContent;
 import static com.xebialabs.restito.semantics.Action.status;
 import static com.xebialabs.restito.semantics.Condition.composite;
+import static com.xebialabs.restito.semantics.Condition.matchesUri;
 import static com.xebialabs.restito.semantics.Condition.method;
 import static com.xebialabs.restito.semantics.Condition.post;
 import static com.xebialabs.restito.semantics.Condition.startsWithUri;
@@ -68,6 +71,47 @@ public class EdcSupport {
                 EDC_API_KEY_HEADER
         ).then(
                 noContent()
+        );
+    }
+
+    public void edcWillReturnContractAgreements() {
+        whenHttp(restitoProvider.stubServer()).match(
+                post("/management/v2/contractagreements/request"),
+                EDC_API_KEY_HEADER
+        ).then(
+                status(HttpStatus.OK_200),
+                restitoProvider.jsonResponseFromFile("stubs/edc/post/data/contractagreements/all_contractagreements_response_200.json")
+        );
+    }
+
+    public void edcWillReturnOnlyOneContractAgreement() {
+        whenHttp(restitoProvider.stubServer()).match(
+                post("/management/v2/contractagreements/request"),
+                EDC_API_KEY_HEADER
+        ).then(
+                status(HttpStatus.OK_200),
+                restitoProvider.jsonResponseFromFile("stubs/edc/post/data/contractagreements/one_contractagreement_response_200.json")
+        );
+    }
+
+    public void edcWillReturnPaginatedContractAgreements() {
+        whenHttp(restitoProvider.stubServer()).match(
+                post("/management/v2/contractagreements/request"),
+                EDC_API_KEY_HEADER
+        ).then(
+                status(HttpStatus.OK_200)
+        ).withSequence(
+                restitoProvider.jsonResponseFromFile("stubs/edc/post/data/contractagreements/first_page_contractagreements_response_200.json"),
+                restitoProvider.jsonResponseFromFile("stubs/edc/post/data/contractagreements/second_page_contractagreements_response_200.json"));
+    }
+
+    public void edcWillReturnContractAgreementNegotiation() {
+        whenHttp(restitoProvider.stubServer()).match(
+                matchesUri(Pattern.compile("/management/v2/contractagreements/[\\w]+/negotiation")),
+                EDC_API_KEY_HEADER
+        ).then(
+                status(HttpStatus.OK_200),
+                restitoProvider.jsonResponseFromFile("stubs/edc/post/data/contractagreements/contractagreement_negotiation_response_200.json")
         );
     }
 
@@ -185,4 +229,6 @@ public class EdcSupport {
                 startsWithUri("/management/v2/contractdefinitions")
         );
     }
+
+
 }